// Capstone Project/src/app/pages/landlord/host-profile-page/screen/types.ts
export enum hostProfileEnum {
  CREATE_DEPOSIT_REQUEST = 'APP/HOST_PROFILE/CREATE_DEPOSIT_REQUEST',
  CREATE_DEPOSIT_SUCCESS = 'APP/HOST_PROFILE/CREATE_DEPOSIT_SUCCESS',
  CREATE_DEPOSIT_FAILURE = 'APP/HOST_PROFILE/CREATE_DEPOSIT_FAILURE',

  GET_TRANSACTION_REQUEST = 'APP/HOST_PROFILE/GET_TRANSACTION_REQUEST',
  GET_TRANSACTION_SUCCESS = 'APP/HOST_PROFILE/GET_TRANSACTION_SUCCESS',
  GET_TRANSACTION_FAILURE = 'APP/HOST_PROFILE/GET_TRANSACTION_FAILURE',

  SEARCH_TRANSACTION_REQUEST = 'APP/HOST_PROFILE/SEARCH_TRANSACTION_REQUEST',
  SEARCH_TRANSACTION_SUCCESS = 'APP/HOST_PROFILE/SEARCH_TRANSACTION_SUCCESS',
  SEARCH_TRANSACTION_FAILURE = 'APP/HOST_PROFILE/SEARCH_TRANSACTION_FAILURE',

  GET_USER_INFO_REQUEST = 'APP/HOST_PROFILE/GET_USER_INFO_REQUEST',
  GET_USER_INFO_SUCCESS = 'APP/HOST_PROFILE/GET_USER_INFO_SUCCESS',
  GET_USER_INFO_FAILURE = 'APP/HOST_PROFILE/GET_USER_INFO_FAILURE',

  GET_USER_INFO_BY_ID_REQUEST = 'APP/HOST_PROFILE/GET_USER_INFO_BY_ID_REQUEST',
  GET_USER_INFO_BY_ID_SUCCESS = 'APP/HOST_PROFILE/GET_USER_INFO_BY_ID_SUCCESS',
  GET_USER_INFO_BY_ID_FAILURE = 'APP/HOST_PROFILE/GET_USER_INFO_BY_ID_FAILURE',

  UPDATE_USER_INFO_REQUEST = 'APP/HOST_PROFILE/UPDATE_USER_INFO_REQUEST',
  UPDATE_USER_INFO_SUCCESS = 'APP/HOST_PROFILE/UPDATE_USER_INFO_SUCCESS',
  UPDATE_USER_INFO_FAILURE = 'APP/HOST_PROFILE/UPDATE_USER_INFO_FAILURE',

  UPDATE_USER_IMAGE_REQUEST = 'APP/HOST_PROFILE/UPDATE_USER_IMAGE_REQUEST',
  UPDATE_USER_IMAGE_SUCCESS = 'APP/HOST_PROFILE/UPDATE_USER_IMAGE_SUCCESS',
  UPDATE_USER_IMAGE_FAILURE = 'APP/HOST_PROFILE/UPDATE_USER_IMAGE_FAILURE',

  CLEAR_HOST_PROFILE_STATE = 'APP/HOST_PROFILE/CLEAR_HOST_PROFILE_STATE',
  CLEAR_MSG_UPDATE = 'APP/HOST_PROFILE/CLEAR_MSG_UPDATE',
}